refactor: role based access to creating team and registering members

+++ src/app/api/team/route.ts
@@ -2,6 +2,7 @@ import { v4 } from 'uuid';
 import { z } from 'zod';
 
 import { teamFormSchema, teamSchemaType } from '@/lib/schemas';
+import { getRole } from '@/lib/server-user';
 import { createSupabaseServerClient } from '@/lib/supabase/server-clients';
 
 export async function POST(req: Request) {
@@ -15,6 +16,7 @@ export async function POST(req: Request) {
       data: { user: serverUser },
     } = await serverSupabase.auth.getUser();
     const user_id = serverUser?.id;
+    const role = await getRole(user_id as string);
     // eslint-disable-next-line unused-imports/no-unused-vars
     const { team_head, members, ...team } = teamData;
     if (user_id) {
@@ -26,14 +28,15 @@ export async function POST(req: Request) {
         projects_done: 0,
         projects_in_progress: 0,
       };
-
-      //first delete team head from members
-      const { error: membersError } = await serverSupabase
-        .from('members')
-        .delete()
-        .eq('member_id', user_id);
-      if (membersError) {
-        throw new Error(membersError.message);
+      if (role === 'TEAM_MEMBER') {
+        //first delete team head from members
+        const { error: membersError } = await serverSupabase
+          .from('members')
+          .delete()
+          .eq('member_id', user_id);
+        if (membersError) {
+          throw new Error(membersError.message);
+        }
       }
 
       members.push(user_id);
@@ -59,6 +62,15 @@ export async function POST(req: Request) {
         .insert([{ ...teamData }]);
       if (teamError) {
         throw new Error('Failed to create team');
+      }
+
+      //update role to TEAM_HEAD
+      const { error: roleError } = await serverSupabase
+        .from('profiles')
+        .update({ role: 'TEAM_HEAD' })
+        .eq('id', user_id);
+      if (roleError) {
+        throw new Error(roleError.message);
       }
 
       return new Response(JSON.stringify(teamData), { status: 200 });
